feature: add support for database upload from the working directory
This PR adds the support for uploading the database dump directly from
the working directory. The user needs to provide the website id where
the database needs to be uploaded and a valid dump file that could be
restored.

In order to make sure the db dump is compatible, create a dump using the
following command:

```
pg_dump --no-privileges --no-acl --no-owner -d cms > cms_dump.sql
```
Now, in order to upload dump, run this command from the directory where
the file exists.

```
divio project --id=38800 push local_db dump_filename tabijn_baken.sql
```

This will upload the file after compressing it and restore it in the
same way it does while uploading db from docker container for the
project.

- Github Issue: fixes #195

Signed-off-by: Vinit Kumar <vinit.kumar@socialschools.nl>