tests: _divio_project: disable stdout capturing during setup command; add more debug output
When running the integration tests, the test suite has to setup a fully
populated divio project, and the success of this stage is crucial for the tests
to succeed. Therefore, the console output of the setup command is helpful for
debugging.

Previously, you could only switch on console output globally, which gets hard
to read since the following tests produce output too.

This patch disables pytests output capturing during the setup stage, adds some
more debug prints, and re enables it afterwards.

Signed-off-by: Florian Scherf <florian.scherf@divio.ch>