Add support for template services during app creation
App templates can be used optionally during app creation and they
include, inter alia, information about the required services. Until now,
only a warning was displayed to the users so as to inform them that they
will have to attach them manually. This patch consist of a solution so
as to inform the user about detected services and to ask permission so
as to add them automatically.