cli: logging: silence azure-storage-blob logging when not in debug mode
azure-storage-blob comes with custom logging handlers which don't respect the
logging base config. This results in ugly formatting issues when using
`app pull media` or `app push media`.

This patch adds a check to the logging setup, if the cli is running in debug
mode, and if not, silences all azure loggers, by setting their log level to
`logging.ERROR`. This way, the azure API gets less noisy, but no important
information is lost. Also it can be re enabled by starting the cli with `-vvv`.

Signed-off-by: Florian Scherf <florian.scherf@divio.ch>