feat: remove pypy tests
I gather not many of our customers are using PyPy, and currently it is a
pain to install the cryptography package (which now requires rust and
openssl libraries to build from source).

This commit removes the systematic testing on pypy, but I am confident
the cli should still work on pypy if needed by a client.