fix: fix KeyError in deployment listing
Recently, we removed the `author` field from the deployment lists view of the
control-panel. This field was never of actual use to anyone because only it
contained the UUID of an Divio account that you couldn't use, nor verify, using
the Divio rest API.

This patch removes the table column `author` from the deployments list command.

Signed-off-by: Florian Scherf <florian.scherf@divio.ch>