cli: setup: fix exception when pulling media
Previously, `pull_media` would just always create a backup for the default
object storage and download it, regardless of whether it existed or contained
anything.
Prior to API v3, this worked without any issues but in v3 we introduced checks
that return an error if no matching service instance is available.

This patch introduces a check to `pull_media` that ensures a matching service
instance is available before trying to backup it.

Signed-off-by: Florian Scherf <florian.scherf@divio.ch>